Migrate TokenCard to shadcn registry (#27)
- Update TokenCard to use components from shadcn registry
- Add shadcn configuration files
- Update installation instructions
