+++ src/app/docs/components/token-card/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import React, { useState, useCallback } from "react";
-import { TokenCard } from "./component";
+import { TokenCard } from "@/components/w3-kit/token-card";
 import { Code, Eye } from "lucide-react";
 import { CodeBlock } from "@/components/docs/codeBlock";
-import { Token } from "../token-list/types";
+import { Token } from "@/components/w3-kit/token-card-types";
 
 // Define the ExtendedToken interface
 interface ExtendedToken extends Token {
@@ -200,7 +200,7 @@ export default function Page() {
                   <p className="text-sm text-gray-600 dark:text-gray-400 mb-4">
                     Run the following command to add the Token Card component to your project:
                   </p>
-                  <CodeBlock code="npx w3-kit@latest add token-card" id="cli" />
+                  <CodeBlock code="npx shadcn@latest add https://w3-kit.com/registry/token-card.json" id="cli" />
                   <p className="text-sm text-gray-600 dark:text-gray-400 mt-4">
                     This will:
                   </p>
